+++ R/combinespecies.r
@@ -1,12 +1,15 @@
-combinespecies<-function(X, min.order = 1, max.order = 3, min.occ = 0, FUN = min, verbose=FALSE, add.names=TRUE, ...) {
+combinespecies<-function(X, min.order = 1, max.order = 3, min.occ = 1, FUN = min, verbose=FALSE, add.names=TRUE, ...) {
 
+  #Turn into a matrix (if not)
+  X = as.matrix(X)
+  
   if(sum(is.na(X))>0) stop("Cannot deal with NA values. Remove and run again.")
   if(min.order<1) stop("min.order must be at least 1.")
   if(min.order>max.order) stop("min.order must be equal or larger than max.order.")
   nsites = nrow(X)
   
   #Get species names
-  spplist = names(X)
+  spplist = colnames(X)
   nspecies = length(spplist)
   if(verbose) cat(paste("Number of species to combine: ",nspecies,"\n", sep=""))
   if(length(spplist)==1) stop("At least two species are necessary.")
@@ -38,18 +41,20 @@ combinespecies<-function(X, min.order = 1, max.order = 3, min.occ = 0, FUN = min
   nc = length(comblistDef)
   if(verbose) cat(paste("Final number of combinations: ",nc,"\n", sep=""))
   
-  C = data.frame(matrix(0,nrow=nc,ncol=nspecies))
-  names(C)<-spplist
-  XC = data.frame(matrix(0, nrow=nrow(X), ncol=nc))
-  row.names(XC) = row.names(X)
+  C = matrix(0,nrow=nc,ncol=nspecies)
+  colnames(C)<-spplist
+  XC = matrix(0, nrow=nrow(X), ncol=nc)
+  rownames(XC) = rownames(X)
   for(i in 1:nc) {
     spvec = as.numeric(comblistDef[[i]])
     XC[,i] <-apply(X[,spvec, drop=FALSE],1,FUN=FUN,...)
     C[i,spvec] = 1
   }
+  cn = 1:ncol(XC)
   if(add.names) {
-    for(i in 1:ncol(XC)) names(XC)[i] = paste(spplist[C[i,]==1], collapse="+")
-  } else names(XC) = 1:ncol(XC)
-  row.names(C) = names(XC)
+    for(i in 1:nc) cn[i] = paste(spplist[C[i,]==1], collapse="+")
+  }
+  colnames(XC) = cn
+  rownames(C) = cn
   return(list(XC = XC, C=C))
 }